[space_map (rust)] Factor out space_map_metadata
