+++ src/cache/check.rs
@@ -12,7 +12,7 @@ use crate::pdata::array_walker::*;
 use crate::pdata::bitset::*;
 use crate::pdata::space_map::*;
 use crate::pdata::space_map_checker::*;
-use crate::pdata::space_map_disk::*;
+use crate::pdata::space_map_common::*;
 use crate::pdata::unpack::unpack;
 use crate::report::*;
