+++ src/pdata/mod.rs
@@ -9,5 +9,7 @@ pub mod btree_merge;
 pub mod btree_walker;
 pub mod space_map;
 pub mod space_map_checker;
+pub mod space_map_common;
 pub mod space_map_disk;
+pub mod space_map_metadata;
 pub mod unpack;
