+++ src/thin/check.rs
@@ -10,7 +10,7 @@ use crate::pdata::btree::{self, *};
 use crate::pdata::btree_walker::*;
 use crate::pdata::space_map::*;
 use crate::pdata::space_map_checker::*;
-use crate::pdata::space_map_disk::*;
+use crate::pdata::space_map_common::*;
 use crate::pdata::unpack::*;
 use crate::report::*;
 use crate::thin::block_time::*;
